+++ tests/Factory/ResponseFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace webignition\Tests\CssValidatorWrapper\Factory;
+namespace webignition\CssValidatorWrapper\Tests\Factory;
 
 use GuzzleHttp\Psr7\Response;
 use Psr\Http\Message\ResponseInterface;
